pyproject.toml
[project]
name = "gymbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "dotenv>=0.9.9",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-telegram-bot>=22.6",
    "python-telegram-bot-calendar>=1.0.5",
    "sqlalchemy>=2.0.46",
]
